requirements.txt
-i http://mirrors.aliyun.com/pypi/simple/ --trusted-host mirrors.aliyun.com
django
easy_thumbnails 
django-pagination 
django-suit 
django-mptt
django-environ
django-secure
django_extensions
django-reversion
django-imagekit
django-import-export
django-simditor

whitenoise

django-debug_toolbar 
drf_generators
click
raven
plan